+++ front/src/types.ts
@@ -3,4 +3,10 @@ type FAQ = {
     pageTitle: string;
 };
 
-export type { FAQ };
+type Page = {
+    lines: Array<{id: string; text: string}>;
+  };
+
+type Lines = Array<{id: string; text: string}>;
+
+export type { FAQ, Page, Lines };
